refactor: move rs-sdk to dash-sdk-bindings
