+++ dash-sdk-android/src/main/rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dash-sdk-bindings"
-version = "0.1.0"
+version = "1.0.0"
 edition = "2021"
 
 [dependencies]
